+++ src/lib.rs
@@ -1,0 +1,52 @@
+pub mod hardware;
+pub mod util;
+
+pub use hardware::sys::System;
+
+pub use ggez::conf::WindowMode;
+pub use ggez::{GameError, GameResult};
+pub use ggez::event::{self, EventHandler};
+pub use ggez::graphics::{self, Color};
+pub use ggez::timer::check_update_time;
+
+pub const DESIRED_FPS: f32 = 50.;
+
+pub struct IbmPc {
+    pub sys: System,
+}
+
+impl IbmPc {
+    pub fn new() -> Self {
+        IbmPc {
+            sys: System::new()
+        }
+    }
+
+    pub fn run_test(&mut self) {
+        loop {
+            self.sys.step(&mut 0u32);
+        }
+    }
+}
+
+impl EventHandler for IbmPc {
+    fn update(&mut self, ctx: &mut ggez::Context) -> Result<(), GameError> {
+        let mut veces = 0;
+
+        while check_update_time(ctx, DESIRED_FPS as u32) {
+            self.sys.update();
+            veces += 1;
+        }
+
+        // println!("{veces} - {}", ggez::timer::fps(ctx));
+
+        Ok(())
+    }
+
+    fn draw(&mut self, ctx: &mut ggez::Context) -> Result<(), GameError> {
+        graphics::clear(ctx, Color::WHITE);
+        // TODO
+
+        graphics::present(ctx)
+    }
+}
