+++ tests/lib.rs
@@ -1,0 +1,8 @@
+use ibm_5150::*;
+
+#[test]
+fn test_main() {
+    let mut sys = IbmPc::new();
+
+    sys.run_test();
+}
